requirements.txt
aiogramaiogram==3.0.0
python-dotenv==1.0.0